+++ clients/mobile/components/tasks/task-filter-sheet.tsx
@@ -1,4 +1,3 @@
-import Feather from "@expo/vector-icons/Feather";
 import { useEffect, useRef, useState } from "react";
 import {
   Animated,
@@ -12,305 +11,12 @@ import {
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
+import { DepartmentSection } from "./department-section";
+import { LocationSection } from "./location-section";
+import { PrioritySection } from "./priority-section";
+import { SortBySection } from "./sort-by-section";
+
 const { height: SCREEN_HEIGHT } = Dimensions.get("window");
-
-// ─── Primitives ────────────────────────────────────────────────────────────────
-
-function CheckedBox() {
-  return (
-    <View
-      style={{
-        width: 16,
-        height: 16,
-        backgroundColor: "#124425",
-        borderRadius: 4,
-        alignItems: "center",
-        justifyContent: "center",
-      }}
-    >
-      <Feather name="check" size={9} color="white" />
-    </View>
-  );
-}
-
-function UncheckedBox() {
-  return (
-    <View
-      style={{
-        width: 16,
-        height: 16,
-        borderWidth: 0.8,
-        borderColor: "#2e2e2e",
-        borderRadius: 4,
-        opacity: 0.5,
-      }}
-    />
-  );
-}
-
-function RadioSelected() {
-  return (
-    <View
-      style={{
-        width: 16,
-        height: 16,
-        borderRadius: 8,
-        borderWidth: 0.8,
-        borderColor: "#124425",
-        alignItems: "center",
-        justifyContent: "center",
-      }}
-    >
-      <View
-        style={{
-          width: 10,
-          height: 10,
-          borderRadius: 5,
-          backgroundColor: "#124425",
-        }}
-      />
-    </View>
-  );
-}
-
-function RadioUnselected() {
-  return (
-    <View
-      style={{
-        width: 16,
-        height: 16,
-        borderRadius: 8,
-        borderWidth: 0.8,
-        borderColor: "#2e2e2e",
-        opacity: 0.4,
-      }}
-    />
-  );
-}
-
-// ─── Row items ─────────────────────────────────────────────────────────────────
-
-function RadioItem({ label, selected }: { label: string; selected: boolean }) {
-  return (
-    <View className="flex-row items-center gap-1">
-      {selected ? <RadioSelected /> : <RadioUnselected />}
-      <Text
-        className="text-[15px] tracking-tight"
-        style={{ color: selected ? "#15502c" : "#040506" }}
-      >
-        {label}
-      </Text>
-    </View>
-  );
-}
-
-function CheckboxItem({ label, checked }: { label: string; checked: boolean }) {
-  return (
-    <View className="flex-row items-center gap-1">
-      {checked ? <CheckedBox /> : <UncheckedBox />}
-      <Text
-        className="text-[15px] tracking-tight"
-        style={{ color: checked ? "#15502c" : "#040506" }}
-      >
-        {label}
-      </Text>
-    </View>
-  );
-}
-
-// ─── Section header ─────────────────────────────────────────────────────────────
-
-type SectionHeaderProps = {
-  label: string;
-  expanded: boolean;
-  onToggle: () => void;
-  icon?: React.ComponentProps<typeof Feather>["name"];
-  customIcon?: React.ReactNode;
-};
-
-function SectionHeader({
-  label,
-  expanded,
-  onToggle,
-  icon,
-  customIcon,
-}: SectionHeaderProps) {
-  return (
-    <Pressable
-      onPress={onToggle}
-      className="flex-row items-center justify-between"
-    >
-      <View className="flex-row items-center gap-1.5">
-        {customIcon ??
-          (icon ? <Feather name={icon} size={14} color="#464646" /> : null)}
-        <Text className="text-[15px] text-[#464646] tracking-tight">
-          {label}
-        </Text>
-      </View>
-      <View className="w-6 h-6 items-center justify-center">
-        <Feather
-          name={expanded ? "chevron-up" : "chevron-down"}
-          size={14}
-          color="#464646"
-        />
-      </View>
-    </Pressable>
-  );
-}
-
-// ─── Sort By ────────────────────────────────────────────────────────────────────
-
-function SortBySection({
-  expanded,
-  onToggle,
-}: {
-  expanded: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <View className="gap-2">
-      <SectionHeader
-        label="Sort By"
-        expanded={expanded}
-        onToggle={onToggle}
-        customIcon={
-          <View style={{ width: 18, height: 16, alignItems: "center" }}>
-            <Feather name="arrow-up" size={9} color="#464646" />
-            <Feather name="arrow-down" size={9} color="#464646" />
-          </View>
-        }
-      />
-      {expanded && (
-        <View className="flex-row justify-between px-1">
-          <RadioItem label="Priority" selected />
-          <RadioItem label="Newest" selected={false} />
-          <RadioItem label="Oldest" selected={false} />
-        </View>
-      )}
-    </View>
-  );
-}
-
-// ─── Priority ───────────────────────────────────────────────────────────────────
-
-function PrioritySection({
-  expanded,
-  onToggle,
-}: {
-  expanded: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <View className="gap-2">
-      <SectionHeader
-        label="Priority"
-        expanded={expanded}
-        onToggle={onToggle}
-        icon="flag"
-      />
-      {expanded && (
-        <View className="flex-row justify-between px-1">
-          <CheckboxItem label="High" checked />
-          <CheckboxItem label="Medium" checked />
-          <CheckboxItem label="Low" checked />
-        </View>
-      )}
-    </View>
-  );
-}
-
-// ─── Department ─────────────────────────────────────────────────────────────────
-
-const DEPT_LEFT = ["Food & Beverage", "Front Office", "Housekeeping"];
-const DEPT_RIGHT = ["Maintenance", "Management", "Security"];
-const DEPT_CHECKED = new Set(["Food & Beverage", "Front Office"]);
-
-function DepartmentSection({
-  expanded,
-  onToggle,
-}: {
-  expanded: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <View className="gap-2">
-      <SectionHeader
-        label="Department"
-        expanded={expanded}
-        onToggle={onToggle}
-        icon="home"
-      />
-      {expanded && (
-        <View className="flex-row justify-between px-1">
-          <View className="gap-1">
-            {DEPT_LEFT.map((d) => (
-              <CheckboxItem key={d} label={d} checked={DEPT_CHECKED.has(d)} />
-            ))}
-          </View>
-          <View className="gap-1">
-            {DEPT_RIGHT.map((d) => (
-              <CheckboxItem key={d} label={d} checked={DEPT_CHECKED.has(d)} />
-            ))}
-          </View>
-        </View>
-      )}
-    </View>
-  );
-}
-
-// ─── Location (Rooms) ────────────────────────────────────────────────────────────
-
-const ROOMS = ["Room 101", "Room 102", "Room 201", "Room 202", "Room 301"];
-
-function LocationSection({
-  expanded,
-  onToggle,
-}: {
-  expanded: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <View className="gap-2">
-      <SectionHeader
-        label="Location"
-        expanded={expanded}
-        onToggle={onToggle}
-        icon="map-pin"
-      />
-      {expanded && (
-        <View
-          style={{
-            borderWidth: 1,
-            borderColor: "#aeaeae",
-            borderRadius: 4,
-            overflow: "hidden",
-          }}
-        >
-          {/* Search bar */}
-          <View
-            className="flex-row items-center gap-2 px-2 pt-2 pb-3"
-            style={{ borderBottomWidth: 1, borderBottomColor: "#e5e9ed" }}
-          >
-            <View className="w-6 h-6 items-center justify-center">
-              <Feather name="search" size={16} color="#464646" />
-            </View>
-            <Text className="text-xs font-bold text-black tracking-tight">
-              Room:
-            </Text>
-          </View>
-          {/* Room list */}
-          {ROOMS.map((room) => (
-            <View key={room} className="px-4 py-2">
-              <Text className="text-xs text-black tracking-tight">{room}</Text>
-            </View>
-          ))}
-        </View>
-      )}
-    </View>
-  );
-}
-
-// ─── Sheet ───────────────────────────────────────────────────────────────────────
 
 type TaskFilterSheetProps = {
   visible: boolean;
@@ -371,7 +77,7 @@ export function TaskFilterSheet({ visible, onClose }: TaskFilterSheetProps) {
           }).start();
         }
       },
-    }),
+    })
   ).current;
 
   return (
